Add div containers to html and flex layouts
Body was flexed in the column direction. Div containers for 1. the
main body and 2. the cards were added to the html file. Main body card
"inner" was flexed and grown to fill up empty space. To get the cards
to space from one another, a gap was added between them after a div was
added to contain them and a width set. Text-wrapping was allowed for
them to sit inline as well as move to the next line.

Additional styling changes were made according to specifications.